The only way to have non-identical Time on GemStone is to change it
MATimeDescriptionTest validates conditions also by validating against
non-identical instance.  The problem is that Pharo and GemStone have
different approaches to Time type.

Pharo has a Time as wrapper over Duration, which enables it to use
aTime copy to create new instance of a Time.

In GemStone the Time is an immediate type.  It gives SmallTime if
possible. The only way to have a non-identical instance is to change
the time.